// types.ts
export interface FractalParams {
  cReal: number;
  cImag: number;
  zoom: number;
  offsetX: number;
  offsetY: number;
  iterations: number;
  isJulia: boolean;
  colorShift: number;
  animateC: boolean;
  animationSpeed: number;
  animMag: number;
  exposure: number;
  foldSpeed: number;
  foldSplits: number;
  animDepth: number;
  yaw: number;
  roll: number;
  warp: number;
  mirrorHorizontal: boolean;
  grayscale: boolean;
  antialiasing: boolean;
}